revert to keep just the typing
