+++ pandas/core/missing.py
@@ -2,7 +2,7 @@
 Routines for filling missing data.
 """
 
-from typing import TYPE_CHECKING, Any, List, Optional, Set, Tuple, Union
+from typing import TYPE_CHECKING, Any, List, Optional, Set, Union
 
 import numpy as np
 
@@ -12,17 +12,14 @@ from pandas.compat._optional import import_optional_dependency
 from pandas.core.dtypes.cast import infer_dtype_from_array
 from pandas.core.dtypes.common import (
     ensure_float64,
-    is_datetime64_any_dtype,
     is_datetime64_dtype,
     is_datetime64tz_dtype,
     is_integer_dtype,
-    is_numeric_dtype,
     is_numeric_v_string_like,
     is_scalar,
     is_timedelta64_dtype,
     needs_i8_conversion,
 )
-from pandas.core.dtypes.generic import ABCMultiIndex
 from pandas.core.dtypes.missing import isna
 
 if TYPE_CHECKING:
@@ -102,14 +99,7 @@ def clean_fill_method(method, allow_nearest=False):
 
 def clean_interp_method(
     method: str, index: "Index", order: Optional[int] = None, **kwargs
-) -> Tuple[str, np.ndarray]:
-    """
-    Validate Index and order keyword for interpolation methods.
-
-    Returns
-    -------
-    tuple of str, np.ndarray
-    """
+) -> str:
     valid = [
         "linear",
         "time",
@@ -134,49 +124,8 @@ def clean_interp_method(
         raise ValueError("You must specify the order of the spline or polynomial.")
     if method not in valid:
         raise ValueError(f"method must be one of {valid}. Got '{method}' instead.")
-    if method in ("krogh", "piecewise_polynomial", "pchip"):
-        if not index.is_monotonic:
-            raise ValueError(
-                f"{method} interpolation requires that the index be monotonic."
-            )
-    elif method == "time":
-        if not getattr(index, "is_all_dates", None):
-            raise ValueError(
-                "time-weighted interpolation only works "
-                "on Series or DataFrames with a DatetimeIndex"
-            )
-        method = "values"
 
-    if method == "linear":
-        xvalues = np.arange(len(index))
-    else:
-        if isinstance(index, ABCMultiIndex):
-            raise ValueError(
-                "Only `method=linear` interpolation is supported on MultiIndexes."
-            )
-
-        methods = {"index", "values", "nearest", "time"}
-        is_numeric_or_datetime = (
-            is_numeric_dtype(index.dtype)
-            or is_datetime64_any_dtype(index.dtype)
-            or is_timedelta64_dtype(index.dtype)
-        )
-        if method not in methods and not is_numeric_or_datetime:
-            raise ValueError(
-                "Index column must be numeric or datetime type when "
-                f"using {method} method other than linear. "
-                "Try setting a numeric or datetime index column before "
-                "interpolating."
-            )
-        if isna(index).any():
-            raise NotImplementedError(
-                "Interpolation with NaNs in the index "
-                "has not been implemented. Try filling "
-                "those NaNs before interpolating."
-            )
-        xvalues = index.values
-
-    return method, xvalues
+    return method
 
 
 def find_valid_index(values, how: str):
@@ -249,6 +198,16 @@ def interpolate_1d(
 
     if valid.all():
         return yvalues
+
+    if method == "time":
+        if not getattr(xvalues, "is_all_dates", None):
+            # if not issubclass(xvalues.dtype.type, np.datetime64):
+            raise ValueError(
+                "time-weighted interpolation only works "
+                "on Series or DataFrames with a "
+                "DatetimeIndex"
+            )
+        method = "values"
 
     valid_limit_directions = ["forward", "backward", "both"]
     limit_direction = limit_direction.lower()
