+++ pandas/core/series.py
@@ -40,7 +40,6 @@ from pandas.util._validators import validate_bool_kwarg, validate_percentile
 from pandas.core.dtypes.cast import (
     convert_dtypes,
     maybe_cast_to_extension_array,
-    maybe_downcast_to_dtype,
     validate_numeric_casting,
 )
 from pandas.core.dtypes.common import (
@@ -93,7 +92,6 @@ from pandas.core.indexes.period import PeriodIndex
 from pandas.core.indexes.timedeltas import TimedeltaIndex
 from pandas.core.indexing import check_bool_indexer
 from pandas.core.internals import SingleBlockManager
-import pandas.core.missing as missing
 from pandas.core.sorting import ensure_key_mapped
 from pandas.core.strings import StringMethods
 from pandas.core.tools.datetimes import to_datetime
@@ -2223,58 +2221,6 @@ Name: Max Speed, dtype: float64
         else:
             # scalar
             return result.iloc[0]
-
-    @doc(NDFrame.interpolate)
-    def interpolate(
-        self,
-        method: str = "linear",
-        axis: Axis = 0,
-        limit: Optional[int] = None,
-        inplace: bool = False,
-        limit_direction: str = "forward",
-        limit_area: Optional[str] = None,
-        downcast: Optional[str] = None,
-        **kwargs,
-    ) -> Optional["Series"]:
-        inplace = validate_bool_kwarg(inplace, "inplace")
-
-        axis = self._get_axis_number(axis)
-
-        # for the methods backfill, bfill, pad, ffill limit_direction and limit_area
-        # are being ignored, see gh-26796 for more information
-        if method in ["backfill", "bfill", "pad", "ffill"]:
-            return self.fillna(
-                method=method,
-                axis=axis,
-                inplace=inplace,
-                limit=limit,
-                downcast=downcast,
-            )
-
-        # validate the interp method and get xvalues
-        method, xvalues = missing.clean_interp_method(method, self.index, **kwargs)
-
-        arr = missing.interpolate_1d(
-            xvalues,
-            self.values,
-            method=method,
-            limit=limit,
-            limit_direction=limit_direction,
-            limit_area=limit_area,
-            bounds_error=False,
-            **kwargs,
-        )
-
-        if downcast is not None:
-            arr = maybe_downcast_to_dtype(arr, dtype=downcast)
-
-        result = self._constructor(arr, index=self.index, fastpath=True)
-
-        if inplace:
-            self._update_inplace(result)
-            return None
-        else:
-            return result.__finalize__(self, method="interpolate")
 
     def corr(self, other, method="pearson", min_periods=None) -> float:
         """
